requirements.txt
openai
python-dotenv
fastapi
uvicorn
gunicorn
python-multipart
gradio
fastapi[standard]
websocket-client
numpy
sounddevice
pydub
soundfile
pymongo